refactor: move auth error to shared ml-5
