+++ apps/backend/src/libs/enums/enums.ts
@@ -1,6 +1,1 @@
-export {
-	APIPath,
-	AppEnvironment,
-	HTTPCode,
-	ServerErrorType,
-} from "@meetlytic/shared";
+export { APIPath, AppEnvironment, ServerErrorType } from "@meetlytic/shared";
