+++ apps/backend/src/modules/auth/libs/enums/enums.ts
@@ -1,1 +1,1 @@
-export { AuthApiPath, AuthStatusMessage } from "@meetlytic/shared";
+export { AuthApiPath } from "@meetlytic/shared";
