--- apps/backend/src/modules/auth/libs/exceptions/auth.error.ts
@@ -1,15 +1,0 @@
-import { HTTPCode } from "~/libs/enums/enums.js";
-import { HTTPError } from "~/libs/exceptions/exceptions.js";
-
-import { AuthStatusMessage } from "../enums/enums.js";
-
-class AuthError extends HTTPError {
-	constructor() {
-		super({
-			message: AuthStatusMessage.INVALID_CREDENTIALS,
-			status: HTTPCode.UNAUTHORIZED,
-		});
-	}
-}
-
-export { AuthError };
