+++ packages/shared/src/libs/exceptions/exceptions.ts
@@ -1,2 +1,3 @@
+export { AuthError } from "./auth-error/auth.error.js";
 export { HTTPError } from "./http-error/http-error.exception.js";
 export { ValidationError } from "./validation-error/validation-error.exception.js";
